:recycle: 여행 스탬프, 여행 태그 orm 분리
