+++ src/database/travel_stamp_orm.py
@@ -5,43 +5,10 @@ from sqlalchemy.sql.sqltypes import Float, Boolean
 from database.base_entity import BaseEntity
 from database.orm import Base
 from database.travel_location_orm import TravelLocation
-from schema.travel_request import TravelLogCreateRequest
 from sqlalchemy.orm import relationship
 
-class TravelLog(Base, BaseEntity):
-    __tablename__ = "travel_log"
-    id = Column(Integer, primary_key=True, autoincrement=True, index=True)
-    title = Column(String(256), nullable=False)
-    subtitle = Column(String(256), nullable=False)
-    summary = Column(Text, nullable=False)
-    theme = Column(String(256), nullable=False)
-    user_id = Column(Integer, ForeignKey("user.id"))
+from database.travel_log_orm import TravelLog
 
-    @classmethod
-    def create(cls, request: TravelLogCreateRequest, user_id: int) -> "TravelLog":
-        return cls(
-            title=request.title,
-            subtitle=request.subtitle,
-            summary=request.summary,
-            theme=request.theme,
-            user_id=user_id,
-        )
-
-
-class TravelLogTag(Base):
-    __tablename__ = "travel_log_tag"
-    id = Column(Integer, primary_key=True, autoincrement=True, index=True)
-    tag = Column(String(256), nullable=False)
-    travel_log_id = Column(Integer, ForeignKey("travel_log.id"))
-    user_id = Column(Integer, ForeignKey("user.id"))
-
-    @classmethod
-    def create(cls, tag: String, travel_log: TravelLog, user_id: int) -> "TravelLogTag":
-        return cls(
-            tag=tag,
-            user_id=user_id,
-            travel_log_id=travel_log.id,
-        )
 
 class TravelStamp(Base, BaseEntity) :
     __tablename__ = "travel_stamp"
