+++ src/repository/travel_log_repository.py
@@ -1,5 +1,10 @@
+from typing import List, Any, Sequence
+
+from sqlalchemy import Row, RowMapping
 from sqlalchemy.orm import Session
-from database.travel_orm import TravelLog
+from sqlalchemy import select, delete
+
+from database.travel_log_orm import TravelLog
 
 
 class TravelLogRepository:
@@ -16,3 +21,6 @@ class TravelLogRepository:
             print(f"Repository에서 에러: {e}")
             self.session.rollback()
             raise
+
+    def get_travel_log_by_user_id(self, user_id: int) -> List[TravelLog]:
+        return self.session.scalars(select(TravelLog).where(TravelLog.user_id == user_id)).unique().all()
